% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose.R
\name{chartWindRose}
\alias{chartWindRose}
\title{Plot wind rose.}
\usage{
chartWindRose(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Get wind data to display the wind rose chart.
}
