% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_data.R
\name{downAWSAggrDataCSV}
\alias{downAWSAggrDataCSV}
\title{Download all variables for one AWS.}
\usage{
downAWSAggrDataCSV(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Download all variables for one AWS.
}
