% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_data_sel.R
\name{downAWSAggrDataSelCSV}
\alias{downAWSAggrDataSelCSV}
\title{Download aggregated data.}
\usage{
downAWSAggrDataSelCSV(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Download aggregated data for a selected stations.
}
