% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes_data.R
\name{downAWSMinOneCSV}
\alias{downAWSMinOneCSV}
\title{Download displayed observation for one AWS.}
\usage{
downAWSMinOneCSV(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Download the time series of the displayed observation for one AWS.
}
