% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_status_data.R
\name{downAWSStatusNoUserReq}
\alias{downAWSStatusNoUserReq}
\title{Download hourly AWS status.}
\usage{
downAWSStatusNoUserReq(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Download hourly AWS status table, no user request.
}
