% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_status_data.R
\name{downAWSStatusUserReqPOST}
\alias{downAWSStatusUserReqPOST}
\title{Download hourly AWS status.}
\usage{
downAWSStatusUserReqPOST(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Download hourly AWS status table, with user request, use POST.
}
