% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_status_data.R
\name{getStatusData}
\alias{getStatusData}
\title{Get AWS Status data.}
\usage{
getStatusData(ltime, varh, aws_dir)
}
\arguments{
\item{ltime}{character, the last time duration to display. Options are, "01h", "03h", "06h", 
"12h", "24h", "02d", "03d", "05d", "01w", "02w", "03w", "01m".}

\item{varh}{character, code variable and height: "codevar_height".}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Get AWS Status data to display on map.
}
