% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_status_data.R
\name{getStatusVariables}
\alias{getStatusVariables}
\title{Get all available variables.}
\usage{
getStatusVariables(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Get all available variables for AWS Data Status.
}
