% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{getWindHeight}
\alias{getWindHeight}
\title{Get AWS Wind heights.}
\usage{
getWindHeight(time_step, aws_dir)
}
\arguments{
\item{time_step}{time step of the data: "minutes" or "hourly".}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Get AWS Wind heights for drop-down select.
}
