% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_contours.R
\name{graphWindContours}
\alias{graphWindContours}
\title{Plot wind frequencies contour.}
\usage{
graphWindContours(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Get wind data to plot the direction frequencies with a contour plot and the speed with a boxplot.
}
