% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gridded_data.R
\name{griddedCountorMap}
\alias{griddedCountorMap}
\title{Plot gridded data, Contour.}
\usage{
griddedCountorMap(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Format gridded data to display on map.
}
