% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_data.R
\name{mapAggrAWSData}
\alias{mapAggrAWSData}
\title{Get AWS aggregated spatial data.}
\usage{
mapAggrAWSData(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Get AWS aggregated spatial data to display on map.
}
