% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minutes_data.R
\name{mapHourAWSData}
\alias{mapHourAWSData}
\title{Get hourly data.}
\usage{
mapHourAWSData(user_req, aws_dir)
}
\arguments{
\item{user_req}{a JSON object containing the user request.}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
JSON object
}
\description{
Get hourly data to display on map.
}
