% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readAWSMetadata}
\alias{readAWSMetadata}
\title{Read AWS metadata for mod_aws module.}
\usage{
readAWSMetadata(time_step, aws_dir)
}
\arguments{
\item{time_step}{time step of the data: "minutes", "hourly" or "daily".}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Read AWS coordinates and parameters for minutes, hourly and daily data.
}
