% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readCoordsPars}
\alias{readCoordsPars}
\title{Read AWS metadata for mod_auth module.}
\usage{
readCoordsPars(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Read AWS coordinates and parameters.
}
