% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readVariableListMap}
\alias{readVariableListMap}
\title{Get variables list.}
\usage{
readVariableListMap(time_step, aws_dir)
}
\arguments{
\item{time_step}{time step of the data: "hourly" or "daily".}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Read the list of variables to display on map.
}
