% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readWindMetadata}
\alias{readWindMetadata}
\title{Get AWS Wind Metadata.}
\usage{
readWindMetadata(time_step, height, aws_dir)
}
\arguments{
\item{time_step}{time step of the data: "minutes" or "hourly".}

\item{height}{wind speed and direction heights above ground, format "<speedHeight>_<directionHeight>".}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Get AWS Wind metadata to display on map.
}
