% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{tableAWSCoords}
\alias{tableAWSCoords}
\title{Get AWS coordinates for one network.}
\usage{
tableAWSCoords(network, aws_dir)
}
\arguments{
\item{network}{the AWS network code;}

\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\value{
a JSON object
}
\description{
Get AWS coordinates for one network to display on table.
}
